pyproject.toml
[tool.poetry]
name = "css-rotator"
version = "0.1.0"
description = ""
authors = ["protogia <goldengrizz01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
